Moving circumsolar and horizon shading models to irradiance folder
