+++ pvfactors/pvarray.py
@@ -5,11 +5,11 @@ from __future__ import division
 from __future__ import print_function
 from pvfactors import PVFactorsArrayUpdateException
 from pvfactors.config import \
-    Y_GROUND, MAX_X_GROUND, MIN_X_GROUND, X_ORIGIN_PVROWS, VIEW_DICT
-from pvfactors.pvcore import (LinePVArray,
-                              find_edge_point,
-                              calculate_circumsolar_shading,
-                              calculate_horizon_band_shading)
+    Y_GROUND, MAX_X_GROUND, MIN_X_GROUND, X_ORIGIN_PVROWS, VIEW_DICT,\
+    DEFAULT_CIRCUMSOLAR_ANGLE, DEFAULT_HORIZON_BAND_ANGLE
+from pvfactors.pvcore import LinePVArray, find_edge_point
+from pvfactors.irradiance.utils import \
+    calculate_circumsolar_shading, calculate_horizon_band_shading
 from pvfactors.pvrow import PVRowLine
 from pvfactors.view_factors import ViewFactorCalculator
 from shapely.geometry import LineString, Point
@@ -29,8 +29,6 @@ LOGGER.setLevel(logging.INFO)
 
 DEFAULT_EDGE_PT_X = X_ORIGIN_PVROWS
 DELTA_MAX_MIN_GROUND_WHEN_TOO_SMALL_BIG = 1
-DEFAULT_CIRCUMSOLAR_ANGLE = 30.
-DEFAULT_HORIZON_BAND_ANGLE = 6.5
 
 
 class ArrayBase(object):
