+++ pvfactors/pvcore.py
@@ -4,7 +4,6 @@ from pvfactors.config import Y_GROUND
 from shapely.geometry import Point
 from pvfactors import PVFactorsError
 import numpy as np
-import math
 import logging
 
 LOGGER = logging.getLogger(__name__)
@@ -12,17 +11,6 @@ LOGGER.setLevel(logging.INFO)
 
 # TODO: hard coding these values is not ideal
 THRESHOLD_EDGE_POINT = 1e3
-RADIUS_CIRCUMSOLAR = 1.
-
-# TODO: hard-coding for registry
-DISTANCE_TOLERANCE = 1e-8
-THRESHOLD_DISTANCE_TOO_CLOSE = 1e-10
-
-# Gaussian shading default parameters: TOTAL_GAUSSIAN_AREA dependent on these
-SIGMA = 1. / np.sqrt(2.)
-MU = 0.
-N_SIGMA = 3.
-GAUSSIAN_DIAMETER_CIRCUMSOLAR = 2. * N_SIGMA * SIGMA
 
 
 class LinePVArray(dict):
@@ -63,206 +51,6 @@ class LinePVArray(dict):
                                  "values are: %s" % (
                                      str(line_type),
                                      str(self._list_line_types)))
-
-
-def calculate_circumsolar_shading(percentage_distance_covered,
-                                  model='uniform_disk'):
-    """Select the model to calculate circumsolar shading based on the current PV
-    array condition.
-
-    Parameters
-    ----------
-    percentage_distance_covered : float
-        this represents how much of the
-        circumsolar diameter is covered by the neighboring row [in %]
-    model : str, optional
-        name of the circumsolar shading model to use:
-        'uniform_disk' and 'gaussian' are the two models currently available
-        (Default value = 'uniform_disk')
-
-    Returns
-    -------
-    float
-        shading percentage of circumsolar area
-
-    """
-    if model == 'uniform_disk':
-        perc_shading = uniform_circumsolar_disk_shading(
-            percentage_distance_covered)
-
-    elif model == 'gaussian':
-        perc_shading = gaussian_shading(percentage_distance_covered)
-
-    else:
-        raise PVFactorsError(
-            'calculate_circumsolar_shading: model does not exist: ' +
-            '%s' % model)
-
-    return perc_shading
-
-
-def integral_default_gaussian(y, x):
-    """Calculate the value of the integral from x to y of the erf function
-
-    Parameters
-    ----------
-    y : float
-        upper limit
-    x : float
-        lower limit
-
-    Returns
-    -------
-    float
-        Calculated value of integral
-
-    """
-    return 0.5 * (math.erf(x) - math.erf(y))
-
-
-def gaussian_shading(percentage_distance_covered):
-    """Calculate circumsolar shading by assuming that the irradiance profile on
-    a 2D section of the circumsolar disk is Gaussian
-
-    Parameters
-    ----------
-    percentage_distance_covered : float
-        [in %], proportion of the
-        circumsolar disk area covered by the neighboring pvrow
-
-    Returns
-    -------
-    float
-        percentage shading in terms of irradiance (using gaussian profile)
-
-    """
-    if percentage_distance_covered < 0.:
-        perc_shading = 0.
-    elif percentage_distance_covered > 100.:
-        perc_shading = 100.
-    else:
-        y = - N_SIGMA * SIGMA
-        x = (y +
-             percentage_distance_covered / 100. *
-             GAUSSIAN_DIAMETER_CIRCUMSOLAR)
-        area = integral_default_gaussian(y, x)
-        total_gaussian_area = integral_default_gaussian(- N_SIGMA * SIGMA,
-                                                        N_SIGMA * SIGMA)
-        perc_shading = area / total_gaussian_area * 100.
-
-    return perc_shading
-
-
-def gaussian(x, mu=0., sigma=1.):
-    """Gaussian density function
-
-    Parameters
-    ----------
-    x : float
-        argument of function
-    mu : float, optional
-        mean of the gaussian (Default value = 0.)
-    sigma : float, optional
-        standard deviation of the gaussian (Default value = 1.)
-
-    Returns
-    -------
-    float
-        value of guassian function at point x
-
-    """
-    return (1. / (sigma * np.sqrt(2. * np.pi)) *
-            np.exp(- 0.5 * np.power((x - mu) / sigma, 2)))
-
-
-def uniform_circumsolar_disk_shading(percentage_distance_covered):
-    """Calculate the percentage shading of circumsolar irradiance. This
-    model considers circumsolar to be a disk, and calculates the
-    percentage shaded based on the percentage "distance covered",
-    which is how much of the disk's diameter is covered by the
-    neighboring object.
-
-    Parameters
-    ----------
-    percentage_distance_covered : float
-        distance covered of circumsolar disk diameter
-        [% - values from 0 to 100]
-
-    Returns
-    -------
-    float
-        value of guassian function at point x
-
-    """
-
-    # Define a circumsolar disk
-    r_circumsolar = RADIUS_CIRCUMSOLAR
-    d_circumsolar = 2 * r_circumsolar
-    area_circumsolar = np.pi * r_circumsolar**2.
-
-    # Calculate circumsolar on case by case
-    distance_covered = percentage_distance_covered / 100. * d_circumsolar
-
-    if distance_covered <= 0:
-        # No shading of circumsolar
-        percent_shading = 0.
-    elif (distance_covered > 0.) & (distance_covered <= r_circumsolar):
-        # Theta is the full circle sector angle (not half) used to
-        # calculate circumsolar shading
-        theta = 2 * np.arccos((r_circumsolar - distance_covered) /
-                              r_circumsolar)  # rad
-        area_circle_sector = 0.5 * r_circumsolar ** 2 * theta
-        area_triangle_sector = 0.5 * r_circumsolar ** 2 * np.sin(theta)
-        area_shaded = area_circle_sector - area_triangle_sector
-        percent_shading = area_shaded / area_circumsolar * 100.
-
-    elif (distance_covered > r_circumsolar) & (distance_covered <
-                                               d_circumsolar):
-        distance_uncovered = d_circumsolar - distance_covered
-
-        # Theta is the full circle sector angle (not half) used to
-        # calculate circumsolar shading
-        theta = 2 * np.arccos((r_circumsolar - distance_uncovered) /
-                              r_circumsolar)  # rad
-        area_circle_sector = 0.5 * r_circumsolar ** 2 * theta
-        area_triangle_sector = 0.5 * r_circumsolar ** 2 * np.sin(theta)
-        area_not_shaded = area_circle_sector - area_triangle_sector
-        percent_shading = (1. - area_not_shaded / area_circumsolar) * 100.
-
-    else:
-        # Total shading of circumsolar: distance_covered >= d_circumsolar
-        percent_shading = 100.
-
-    return percent_shading
-
-
-def calculate_horizon_band_shading(shading_angle, horizon_band_angle):
-    """Calculate the percentage shading of the horizon band; which is just the
-    proportion of what's masked by the neighboring row since we assume
-    uniform luminance values for the band
-
-    Parameters
-    ----------
-    shading_angle : float
-        the elevation angle to use for shading
-    horizon_band_angle : float
-        the total elevation angle of the horizon band
-
-    Returns
-    -------
-    float
-        shading percentage of horizon band
-
-    """
-    percent_shading = 0.
-    if shading_angle >= horizon_band_angle:
-        percent_shading = 100.
-    elif (shading_angle >= 0) | (shading_angle < horizon_band_angle):
-        percent_shading = 100. * shading_angle / horizon_band_angle
-    else:
-        # No shading
-        pass
-    return percent_shading
 
 
 def find_edge_point(b1_pvrow, b2_pvrow):
